Limit remote path fix to plugins

Copilot-Session: a709d3d4-1ecd-44ce-89bc-c343a6e47285
