+++ extensions/copilot/src/extension/chatSessions/claude/node/test/skillConfigLocations.spec.ts
@@ -12,7 +12,6 @@ import { NullNativeEnvService } from '../../../../../platform/env/common/nullEnv
 import { IWorkspaceService } from '../../../../../platform/workspace/common/workspaceService';
 import { Event } from '../../../../../util/vs/base/common/event';
 import { DisposableStore } from '../../../../../util/vs/base/common/lifecycle';
-import { isWindows } from '../../../../../util/vs/base/common/platform';
 import { URI } from '../../../../../util/vs/base/common/uri';
 import { createExtensionUnitTestingServices } from '../../../../test/node/services';
 import { resolveSkillConfigLocations } from '../../../common/skillConfigLocations';
@@ -90,53 +89,6 @@ describe('resolveSkillConfigLocations', () => {
 		expect(result[0].path).toBe('/absolute/skills/path');
 	});
 
-	it.skipIf(!isWindows)('preserves UNC authority for local absolute paths', () => {
-		expect(resolve({
-			configLocations: { '\\\\server\\share\\skills': true },
-			userHome: URI.file('C:\\Users\\user'),
-		})).toEqual([URI.file('\\\\server\\share\\skills')]);
-	});
-
-	it('uses the remote workspace authority for absolute and tilde paths when user home is local', () => {
-		const result = resolve({
-			configLocations: {
-				'/opt/skills': true,
-				'~': true,
-				'~/skills': true,
-			},
-			userHome: URI.file('/home/user'),
-			workspaceFolders: [URI.from({ scheme: 'vscode-remote', authority: 'wsl+ubuntu', path: '/workspace' })],
-		});
-
-		expect(result.map(uri => uri.toString())).toEqual([
-			'vscode-remote://wsl%2Bubuntu/opt/skills',
-			'vscode-remote://wsl%2Bubuntu/home/user',
-			'vscode-remote://wsl%2Bubuntu/home/user/skills',
-		]);
-	});
-
-	it.skipIf(!isWindows)('uses Windows target paths and preserves UNC shares for remote workspaces', () => {
-		const userHome = URI.file('C:\\Users\\user');
-		const workspaceFolder = URI.from({ scheme: 'vscode-remote', authority: 'ssh-remote+windows', path: '/C:/workspace' });
-		const filePath = URI.file('C:\\skills');
-		const tildePath = URI.file(`${userHome.path}\\skills`);
-		const uncPath = URI.file('\\\\server\\share\\skills');
-
-		expect(resolve({
-			configLocations: {
-				'C:\\skills': true,
-				'~\\skills': true,
-				'\\\\server\\share\\skills': true,
-			},
-			userHome,
-			workspaceFolders: [workspaceFolder],
-		})).toEqual([
-			workspaceFolder.with({ path: filePath.path }),
-			workspaceFolder.with({ path: tildePath.path }),
-			workspaceFolder.with({ path: `//${uncPath.authority}${uncPath.path}` }),
-		]);
-	});
-
 	it('joins relative paths to each workspace folder', () => {
 		const result = resolve({
 			configLocations: { 'relative/skills': true },
@@ -145,13 +97,6 @@ describe('resolveSkillConfigLocations', () => {
 		expect(result).toHaveLength(2);
 		expect(result[0].path).toBe('/workspace1/relative/skills');
 		expect(result[1].path).toBe('/workspace2/relative/skills');
-	});
-
-	it('treats invalid tilde-prefixed locations as workspace relative', () => {
-		expect(resolve({
-			configLocations: { '~shared/skills': true },
-			workspaceFolders: [URI.file('/workspace')],
-		})).toEqual([URI.file('/workspace/~shared/skills')]);
 	});
 
 	it('ignores config entries with value !== true', () => {
